Рефакторит переключение периода показа статистики.
